refactor(boxutils): make read/write b3d wrappers of canonical io.model
- read_b3d_h5 delegates to io.model._raw_read_h5 and converts
  GeometryContract back to dict for legacy caller compatibility
- write_b3d_h5 delegates to io.model._raw_write_h5
- io.model gains _raw_read_h5/_raw_write_h5 internal helpers to
  avoid circular dependency with boxutils callers
- All 26 IO/thin/CLI tests passing
