+++ pyampp/gxbox/boxutils.py
@@ -1277,53 +1277,19 @@ def read_b3d_h5(filename):
         by_cor = b3dbox['corona']['by']
         bz_cor = b3dbox['corona']['bz']
     """
-    def _read_h5_node(node):
-        if isinstance(node, h5py.Group):
-            out = {}
-            for key in node.keys():
-                out[key] = _read_h5_node(node[key])
-            return out
-        if node.shape == ():
-            return node[()]
-        return node[:]
+    from pyampp.geometry.contract import GeometryContract
+    from pyampp.io import load_model_from_h5
 
-    box_b3d = {}
-    with h5py.File(filename, 'r') as hdf_file:
-        for model_type in hdf_file.keys():
-            group = hdf_file[model_type]
-            target_type = model_type
-            model_attr = None
-            if model_type in ("nlfff", "pot", "potential", "bounds"):
-                target_type = "corona"
-                if model_type == "potential":
-                    model_attr = "pot"
-                elif model_type in ("bounds",):
-                    model_attr = "bnd"
-                else:
-                    model_attr = model_type
-            if target_type not in box_b3d:
-                box_b3d[target_type] = {}
-            component_names = list(group.keys())
-            if target_type == "refmaps":
-                def _refmap_sort_key(name: str):
-                    obj = group[name]
-                    if isinstance(obj, h5py.Group) and "order_index" in obj.attrs:
-                        return (0, int(obj.attrs["order_index"]))
-                    return (1, name)
-
-                component_names = sorted(component_names, key=_refmap_sort_key)
-            for component in component_names:
-                ds = group[component]
-                box_b3d[target_type][component] = _read_h5_node(ds)
-            if len(group.attrs.keys()) > 0 or model_attr is not None:
-                attrs = dict(group.attrs)
-                if model_attr is not None and "model_type" not in attrs:
-                    attrs["model_type"] = model_attr
-                if "attrs" in box_b3d[target_type]:
-                    box_b3d[target_type]["attrs"].update(attrs)
-                else:
-                    box_b3d[target_type]["attrs"] = attrs
-    return box_b3d
+    model = load_model_from_h5(filename)
+    metadata = model.get("metadata") if isinstance(model, dict) else None
+    if isinstance(metadata, dict):
+        contract = metadata.get("geometry_contract")
+        if isinstance(contract, GeometryContract):
+            metadata = dict(metadata)
+            metadata["geometry_contract"] = contract.to_dict()
+            model = dict(model)
+            model["metadata"] = metadata
+    return model
 
 def write_b3d_h5(filename, box_b3d):
     """
@@ -1335,100 +1301,9 @@ def write_b3d_h5(filename, box_b3d):
         A dictionary containing the B3D data to be written.
     """
 
-    def _encode_dataset_value(value):
-        if isinstance(value, (str, bytes, np.bytes_)):
-            return np.bytes_(value)
-        if value is None:
-            return np.bytes_("")
+    from pyampp.io import save_model_to_h5
 
-        arr = np.asarray(value)
-        if arr.dtype.kind != "O":
-            return value
-
-        if arr.shape == ():
-            scalar = arr.item()
-            if isinstance(scalar, (str, bytes, np.bytes_)):
-                return np.bytes_(scalar)
-            if scalar is None:
-                return np.bytes_("")
-            try:
-                return np.asarray(scalar)
-            except Exception:
-                return np.bytes_(str(scalar))
-
-        # Object arrays are serialized as fixed-width bytes arrays for HDF5 safety.
-        as_text = np.vectorize(lambda x: "" if x is None else str(x), otypes=[str])(arr)
-        return as_text.astype("S")
-
-    def _set_group_attrs(group, attrs):
-        if not isinstance(attrs, dict):
-            return
-        for key, value in attrs.items():
-            group.attrs[key] = _encode_dataset_value(value)
-
-    def _write_node(group, key, value, *, target_type=None):
-        if isinstance(value, dict):
-            sub = group.create_group(key)
-            attrs = value.get("attrs", {}) if isinstance(value.get("attrs"), dict) else {}
-            for sub_key, sub_val in value.items():
-                if sub_key == "attrs":
-                    continue
-                _write_node(sub, sub_key, sub_val, target_type=target_type)
-            _set_group_attrs(sub, attrs)
-            return
-
-        if target_type in ("chromo", "lines") and key == "voxel_status":
-            group.create_dataset(key, data=np.asarray(value, dtype=np.uint8))
-            return
-        group.create_dataset(key, data=_encode_dataset_value(value))
-
-    with h5py.File(filename, 'w') as hdf_file:
-        for model_type, components in box_b3d.items():
-            if components is None:
-                print(f"Warning: {model_type} components are None, skipping.")
-                continue
-            if model_type == "metadata":
-                group = hdf_file.create_group("metadata")
-                for key, value in components.items():
-                    _write_node(group, key, value, target_type="metadata")
-                continue
-            target_type = model_type
-            attrs = components.get("attrs", {}) if isinstance(components, dict) else {}
-            if model_type in ("nlfff", "pot", "potential", "bounds"):
-                target_type = "corona"
-                if "model_type" not in attrs:
-                    attrs = dict(attrs)
-                    if model_type == "potential":
-                        attrs["model_type"] = "pot"
-                    elif model_type == "bounds":
-                        attrs["model_type"] = "bnd"
-                    else:
-                        attrs["model_type"] = model_type
-            if target_type in hdf_file:
-                # Avoid collisions if multiple legacy groups map to corona
-                continue
-            if target_type == "refmaps":
-                group = hdf_file.create_group(target_type, track_order=True)
-            else:
-                group = hdf_file.create_group(target_type)
-            refmap_idx = 0
-            for component, data in components.items():
-                if component == "attrs":
-                    continue
-                if target_type == "refmaps" and isinstance(data, dict):
-                    sub = group.create_group(component, track_order=True)
-                    sub.attrs["order_index"] = np.int64(refmap_idx)
-                    refmap_idx += 1
-                    sub_attrs = data.get("attrs", {}) if isinstance(data.get("attrs"), dict) else {}
-                    for sub_key, sub_val in data.items():
-                        if sub_key == "attrs":
-                            continue
-                        _write_node(sub, sub_key, sub_val, target_type=target_type)
-                    _set_group_attrs(sub, sub_attrs)
-                else:
-                    _write_node(group, component, data, target_type=target_type)
-            if attrs:
-                _set_group_attrs(group, attrs)
+    return save_model_to_h5(box_b3d, filename)
 
 
 def update_line_seeds_h5(filename, line_seeds):
